修改节假日加班值不对

+++ project_manage/src/main/java/com/primeton/manage/employee/dao/AttendanceRepository.java
@@ -14,7 +14,7 @@ public interface AttendanceRepository extends JpaRepository<AttendanceRecord, Lo
 	public Double findWorkOverTime(@Param("empId") Long empId, @Param("startDate") String startDate,@Param("endDate") String endDate);
 
 	@Query(value = "SELECT COALESCE(sum(work_outtime),0) FROM attendance_record WHERE emp_id=:empId and attendance_date between :startDate and :endDate AND flag=5 AND work_outtime > 8",nativeQuery = true)
-	public Integer findHolidayWorkingHours(@Param("empId") Long empId, @Param("startDate") String startDate,@Param("endDate") String endDate);
+	public Double findHolidayWorkingHours(@Param("empId") Long empId, @Param("startDate") String startDate,@Param("endDate") String endDate);
 
 	@Query(value = "SELECT count(1) FROM attendance_record WHERE emp_id=:empId and attendance_date between :startDate and :endDate AND flag=5 AND work_outtime BETWEEN 4 AND 8",nativeQuery = true)
 	public Integer findHolidayWorkingDays(@Param("empId") Long empId, @Param("startDate") String startDate,@Param("endDate") String endDate);
